Renames 'context'-prefixed files to be 'builder' in preparation for a new module called 'context'
